pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "odonto-agenda-ai"
version = "0.1.0"
description = "MVP de agenda odontologica con arquitectura modular por agentes."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "alembic>=1.14.0",
  "fastapi>=0.115.0",
  "jinja2>=3.1.0",
  "psycopg[binary]>=3.2.0",
  "PyJWT>=2.10.0",
  "python-multipart>=0.0.20",
  "pydantic-settings>=2.6.0",
  "sqlalchemy>=2.0.36",
  "email-validator>=2.2.0",
  "uvicorn[standard]>=0.32.0",
  "openai>=1.68.0",
]

[project.optional-dependencies]
dev = [
  "httpx>=0.28.0",
  "pytest>=8.3.0",
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
addopts = "-q"
